requirements.txt
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
emoji==2.8.0
fastapi==0.109.0
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
huggingface-hub==0.36.0
idna==3.11
joblib==1.5.2
motor==3.3.2
nltk==3.8.1
numpy<2
oauthlib==3.3.1
packaging==25.0
pandas==2.1.4
pymongo==4.6.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-oauthlib==1.4.0
safetensors==0.7.0
six==1.17.0
starlette==0.35.0
tokenizers==0.19.1
torch==2.2.0+cpu
tqdm==4.67.1
transformers==4.40.0
tweepy==4.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn[standard]==0.27.0
pytest==7.4.2